refactor(llm): extract EmbeddingMixin + token-budget helpers (Tier-2.5 Task 2)
Move the three self-bound embedding methods (_resolve_default_embedding_model,
get_embedding, get_embeddings) into EmbeddingMixin and the four stateless
token-budget helpers (+ _TRUNCATION_WARNED_MODELS + budget constants) to
_litellm_embedding.py; compose EmbeddingMixin into LiteLLMClient. The mixin
self-types foreign members (config, _default_embedding_model, _resolve_api_key)
via per-mixin TYPE_CHECKING stubs (Tier-1b idiom), never class-level defaults.

SINK-1 (patch-where-used): repoint every embedding patch namespace from the old
litellm_client namespace to _litellm_embedding — resolve_model_name (the :717
embedding-path site), get_service_embeddings, should_use_embedding_service,
_is_chromadb_importable, NomicEmbedder, LocalEmbedder(.get), and
litellm.get_model_info — across test_litellm_client_unit,
test_embedding_service_provider, and test_local_embedding_provider (both the
string-target and setattr(module,...) forms). The suites' hard mock assertions
(assert_not_called / assert_called_once / cache-count == 1) confirm each
repointed mock fires — a no-op patch hitting the real embedder/137M model would
fail loud.

SINK-2 (identity): the facade re-exports _TRUNCATION_WARNED_MODELS + the three
test-imported budget functions by import binding; litellm.embedding stays a
shared-module attr so the global mock still intercepts. Surface guard's
_TRUNCATION_WARNED_MODELS identity assert now enforces object identity.
