+++ reflexio/server/llm/litellm_client.py
@@ -16,14 +16,28 @@ from functools import lru_cache
 from typing import Any
 
 import litellm
-import tiktoken
 from pydantic import BaseModel
 
 from reflexio.models.config_schema import APIKeyConfig
+from reflexio.server.llm._litellm_embedding import (
+    _TRUNCATION_WARNED_MODELS as _TRUNCATION_WARNED_MODELS,
+)
 
 # Identity-preserving re-exports (SINK-2): every moved name is re-bound here by
 # import, never redefined, so ``from ...litellm_client import <name>`` keeps
 # resolving the SAME object/class the moved code uses and tests touch.
+from reflexio.server.llm._litellm_embedding import (
+    EmbeddingMixin,
+)
+from reflexio.server.llm._litellm_embedding import (
+    _get_embedding_encoding as _get_embedding_encoding,
+)
+from reflexio.server.llm._litellm_embedding import (
+    _get_embedding_limit as _get_embedding_limit,
+)
+from reflexio.server.llm._litellm_embedding import (
+    _truncate_for_embedding as _truncate_for_embedding,
+)
 from reflexio.server.llm._litellm_json_extraction import (
     _extract_json_from_string,
     _looks_truncated_json,
@@ -73,26 +87,8 @@ from reflexio.server.llm.model_defaults import ModelRole, resolve_model_name
 from reflexio.server.llm.providers.claude_code_provider import (
     register_if_enabled as _register_claude_code,
 )
-from reflexio.server.llm.providers.embedding_service_provider import (
-    EmbeddingUnavailableError,
-    embedding_provider_mode,
-    get_service_embeddings,
-    should_use_embedding_service,
-)
-from reflexio.server.llm.providers.local_embedding_provider import (
-    LocalEmbedder,
-)
-from reflexio.server.llm.providers.local_embedding_provider import (
-    is_chromadb_importable as _is_chromadb_importable,
-)
 from reflexio.server.llm.providers.local_embedding_provider import (
     register_if_chromadb_available as _register_local_embedder,
-)
-from reflexio.server.llm.providers.nomic_embedding_provider import (
-    NomicEmbedder,
-)
-from reflexio.server.llm.providers.nomic_embedding_provider import (
-    is_nomic_model as _is_nomic_model,
 )
 from reflexio.server.llm.providers.nomic_embedding_provider import (
     register_if_enabled as _register_nomic_embedder,
@@ -124,146 +120,6 @@ __all__ = [
 
 _LOGGER = logging.getLogger(__name__)
 
-# OpenAI's documented max input length for text-embedding-3-* and ada-002 is
-# 8191 tokens. Used as the fallback limit only when a model's name looks
-# OpenAI-family but litellm's registry has no entry for it.
-_OPENAI_EMBEDDING_FALLBACK_MAX_TOKENS = 8191
-
-# Models whose truncation warning has already been emitted this process. Keeps
-# batch backfills of millions of long docs from flooding logs — the first hit
-# per model goes to WARNING, everything after to DEBUG.
-_TRUNCATION_WARNED_MODELS: set[str] = set()
-
-# Model-name prefixes that route through OpenAI's embedding API (and therefore
-# share the 8191-token cap). Anything that does not start with one of these is
-# treated as "unknown provider" when litellm has no registry entry.
-_OPENAI_EMBEDDING_FAMILY_PREFIXES = ("text-embedding-", "openai/", "azure/")
-
-
-@lru_cache(maxsize=32)
-def _get_embedding_limit(model: str) -> int | None:
-    """
-    Resolve the maximum input token count for an embedding model.
-
-    Consults ``litellm.get_model_info`` first so provider-specific caps are
-    respected (OpenAI ~8191, Cohere 512, Voyage 32000, etc.). When litellm has
-    no entry for the model, falls back to the OpenAI 8191 cap only when the
-    model name looks OpenAI-family; otherwise returns ``None`` to disable
-    truncation for unknown providers (safer than over-truncating their input).
-
-    Args:
-        model (str): Embedding model name (e.g. 'text-embedding-3-small',
-            'cohere/embed-english-v3.0').
-
-    Returns:
-        int | None: Maximum input tokens, or ``None`` when the limit is unknown
-            and no safe fallback applies.
-    """
-    try:
-        info = litellm.get_model_info(model)
-    except Exception:
-        info = None
-    if info and info.get("mode") == "embedding":
-        max_tokens = info.get("max_input_tokens")
-        if isinstance(max_tokens, int) and max_tokens > 0:
-            return max_tokens
-    if model.startswith(_OPENAI_EMBEDDING_FAMILY_PREFIXES):
-        return _OPENAI_EMBEDDING_FALLBACK_MAX_TOKENS
-    return None
-
-
-@lru_cache(maxsize=16)
-def _get_embedding_encoding(model: str) -> tiktoken.Encoding:
-    """
-    Return the tiktoken encoding for an embedding model, falling back to cl100k_base.
-
-    For non-OpenAI providers tiktoken does not know the real tokenizer, so the
-    cl100k_base fallback is an approximate proxy for token counting. That is
-    acceptable here because we truncate toward the provider's cap with the
-    proxy, which tends to over-truncate by a small fraction rather than under-
-    truncate and cause upstream 400s.
-
-    Args:
-        model (str): Embedding model name (e.g. 'text-embedding-3-small').
-
-    Returns:
-        tiktoken.Encoding: Encoder to use for token counting and truncation.
-    """
-    try:
-        return tiktoken.encoding_for_model(model)
-    except KeyError:
-        return tiktoken.get_encoding("cl100k_base")
-
-
-def _reject_cloud_mode(embedding_model: str, mode: str) -> None:
-    """
-    Raise when a local-only embedding model is configured for cloud mode.
-
-    Args:
-        embedding_model (str): The resolved embedding model name.
-        mode (str): The resolved embedding provider mode.
-
-    Raises:
-        EmbeddingUnavailableError: If ``mode`` is ``"cloud"``.
-    """
-    if mode == "cloud":
-        raise EmbeddingUnavailableError(
-            f"Local embedding model {embedding_model!r} cannot use cloud mode"
-        )
-
-
-def _truncate_for_embedding(
-    text: str, model: str, max_tokens: int | None = None
-) -> str:
-    """
-    Truncate a string so its token count fits within an embedding model's input limit.
-
-    The token budget is auto-resolved from ``_get_embedding_limit`` by default.
-    When the model has no known limit (unknown provider not in litellm's
-    registry and not OpenAI-family), returns the text unchanged — over-
-    truncating an unknown provider's input is worse than passing it through
-    and letting the provider's own error surface.
-
-    Args:
-        text (str): Raw input text.
-        model (str): Embedding model name, used to pick the tokenizer and the
-            per-provider token cap.
-        max_tokens (int | None): Override for the resolved budget. Primarily
-            used by tests to exercise the truncation path on short strings;
-            leave as ``None`` in production callers.
-
-    Returns:
-        str: Original text if it already fits (or the model has no known
-            limit), otherwise a token-bounded prefix.
-    """
-    if not text:
-        return text
-    if max_tokens is None:
-        max_tokens = _get_embedding_limit(model)
-    if max_tokens is None:
-        return text
-    encoding = _get_embedding_encoding(model)
-    tokens = encoding.encode(text, disallowed_special=())
-    if len(tokens) <= max_tokens:
-        return text
-    if model in _TRUNCATION_WARNED_MODELS:
-        _LOGGER.debug(
-            "Truncating embedding input from %d to %d tokens for model %s",
-            len(tokens),
-            max_tokens,
-            model,
-        )
-    else:
-        _TRUNCATION_WARNED_MODELS.add(model)
-        _LOGGER.warning(
-            "Truncating embedding input from %d to %d tokens for model %s "
-            "(further occurrences will be logged at DEBUG)",
-            len(tokens),
-            max_tokens,
-            model,
-        )
-    return encoding.decode(tokens[:max_tokens])
-
 
 # Per-model provider-timeout floors. Values are floors, not overrides: the
 # effective timeout is max(configured, floor), and an explicit per-call timeout
@@ -280,7 +136,7 @@ _MODEL_TIMEOUT_FLOOR_SECONDS: dict[str, int] = {
 }
 
 
-class LiteLLMClient:
+class LiteLLMClient(EmbeddingMixin):
     """
     Unified LLM client using LiteLLM for multi-provider support.
 
@@ -557,210 +413,6 @@ class LiteLLMClient:
             kwargs["fallback_models"] = fallback_models
 
         return self._make_request(final_messages, **kwargs)
-
-    def _resolve_default_embedding_model(self) -> str:
-        """
-        Resolve the embedding model to use when callers do not specify one.
-
-        Routes through the same auto-detection chain as the rest of reflexio
-        (``resolve_model_name`` for ``ModelRole.EMBEDDING``) so a session that
-        has the local ONNX embedder enabled — or any non-OpenAI provider —
-        does not silently fall back to ``text-embedding-3-small`` and produce
-        OpenAI 401s. Higher-precedence org config and site-var overrides are
-        the caller's responsibility to resolve and pass via ``model=``; this
-        helper handles only the auto-detect tier.
-
-        Returns:
-            str: The auto-detected embedding model name (cached after first call).
-
-        Raises:
-            RuntimeError: Propagated from ``resolve_model_name`` when no
-                embedding-capable provider is available.
-        """
-        if self._default_embedding_model is None:
-            self._default_embedding_model = resolve_model_name(
-                ModelRole.EMBEDDING,
-                api_key_config=self.config.api_key_config,
-            )
-        return self._default_embedding_model
-
-    def get_embedding(
-        self, text: str, model: str | None = None, dimensions: int | None = None
-    ) -> list[float]:
-        """
-        Get embedding vector for the given text.
-
-        Args:
-            text: The text to get embedding for.
-            model: Optional embedding model. When omitted, the model is
-                auto-detected via ``resolve_model_name(ModelRole.EMBEDDING)``
-                so callers inherit the local-embedder gate and any non-OpenAI
-                provider configured for this client.
-            dimensions: Optional number of dimensions for the embedding vector.
-
-        Returns:
-            List of floats representing the embedding vector.
-
-        Raises:
-            LiteLLMClientError: If embedding generation fails.
-        """
-        embedding_model = model or self._resolve_default_embedding_model()
-        mode = embedding_provider_mode(embedding_model)
-        if mode == "off":
-            raise EmbeddingUnavailableError("Embedding provider is disabled")
-        if should_use_embedding_service(embedding_model):
-            return get_service_embeddings(
-                [text], model=embedding_model, dimensions=dimensions
-            )[0]
-
-        # local/nomic-embed-* must stay on the Nomic provider (137M params,
-        # 768d Matryoshka-truncated to 512). Falling through to MiniLM would
-        # mix embedding models inside existing vector stores.
-        if _is_nomic_model(embedding_model):
-            _reject_cloud_mode(embedding_model, mode)
-            try:
-                return NomicEmbedder.get().embed([text])[0]
-            except Exception as e:
-                raise LiteLLMClientError(
-                    f"Nomic embedding generation failed: {str(e)}"
-                ) from e
-
-        # local/* models route through the in-process ONNX embedder — no
-        # network call, no litellm API, no tiktoken truncation (the embedder
-        # applies its own token cap). The dispatch is gated solely on
-        # ``chromadb`` being importable; the env-var opt-in (claude-smart's
-        # ``CLAUDE_SMART_USE_LOCAL_EMBEDDING``) is enforced earlier in the
-        # auto-detection layer (see ``model_defaults._auto_detect_model``).
-        if embedding_model.startswith("local/"):
-            _reject_cloud_mode(embedding_model, mode)
-            if not _is_chromadb_importable():
-                raise LiteLLMClientError(
-                    f"Embedding model {embedding_model!r} requires chromadb. "
-                    "Run `pip install chromadb`."
-                )
-            try:
-                return LocalEmbedder.get().embed([text])[0]
-            except Exception as e:
-                raise LiteLLMClientError(
-                    f"Local embedding generation failed: {str(e)}"
-                ) from e
-
-        text = _truncate_for_embedding(text, embedding_model)
-
-        try:
-            params = {"model": embedding_model, "input": [text]}
-            if dimensions:
-                params["dimensions"] = dimensions
-
-            # Resolve and add API key configuration if provided (overrides env vars)
-            api_key, api_base, api_version = self._resolve_api_key(
-                embedding_model, for_embedding=True
-            )
-            if api_key:
-                params["api_key"] = api_key
-            if api_base:
-                params["api_base"] = api_base
-            if api_version:
-                params["api_version"] = api_version
-
-            response = litellm.embedding(
-                **params,
-                timeout=self.config.timeout,
-                num_retries=self.config.max_retries,
-            )
-            return response.data[0]["embedding"]
-        except Exception as e:
-            raise LiteLLMClientError(f"Embedding generation failed: {str(e)}") from e
-
-    def get_embeddings(
-        self,
-        texts: list[str],
-        model: str | None = None,
-        dimensions: int | None = None,
-    ) -> list[list[float]]:
-        """
-        Get embedding vectors for multiple texts in a single API call.
-
-        Args:
-            texts: List of texts to get embeddings for.
-            model: Optional embedding model. When omitted, the model is
-                auto-detected via ``resolve_model_name(ModelRole.EMBEDDING)``
-                so callers inherit the local-embedder gate and any non-OpenAI
-                provider configured for this client.
-            dimensions: Optional number of dimensions for the embedding vectors.
-
-        Returns:
-            List of embedding vectors, one per input text, in the same order as input.
-
-        Raises:
-            LiteLLMClientError: If embedding generation fails.
-        """
-        if not texts:
-            return []
-
-        embedding_model = model or self._resolve_default_embedding_model()
-        mode = embedding_provider_mode(embedding_model)
-        if mode == "off":
-            raise EmbeddingUnavailableError("Embedding provider is disabled")
-        if should_use_embedding_service(embedding_model):
-            return get_service_embeddings(
-                list(texts), model=embedding_model, dimensions=dimensions
-            )
-
-        # See matching short-circuits in get_embedding above.
-        if _is_nomic_model(embedding_model):
-            _reject_cloud_mode(embedding_model, mode)
-            try:
-                return NomicEmbedder.get().embed(list(texts))
-            except Exception as e:
-                raise LiteLLMClientError(
-                    f"Nomic batch embedding generation failed: {str(e)}"
-                ) from e
-
-        if embedding_model.startswith("local/"):
-            _reject_cloud_mode(embedding_model, mode)
-            if not _is_chromadb_importable():
-                raise LiteLLMClientError(
-                    f"Embedding model {embedding_model!r} requires chromadb. "
-                    "Run `pip install chromadb`."
-                )
-            try:
-                return LocalEmbedder.get().embed(list(texts))
-            except Exception as e:
-                raise LiteLLMClientError(
-                    f"Local batch embedding generation failed: {str(e)}"
-                ) from e
-
-        texts = [_truncate_for_embedding(t, embedding_model) for t in texts]
-
-        try:
-            params = {"model": embedding_model, "input": texts}
-            if dimensions:
-                params["dimensions"] = dimensions
-
-            # Resolve and add API key configuration if provided (overrides env vars)
-            api_key, api_base, api_version = self._resolve_api_key(
-                embedding_model, for_embedding=True
-            )
-            if api_key:
-                params["api_key"] = api_key
-            if api_base:
-                params["api_base"] = api_base
-            if api_version:
-                params["api_version"] = api_version
-
-            response = litellm.embedding(
-                **params,
-                timeout=self.config.timeout,
-                num_retries=self.config.max_retries,
-            )
-            # Response data may not be in order, sort by index to ensure correct ordering
-            sorted_data = sorted(response.data, key=lambda x: x["index"])
-            return [item["embedding"] for item in sorted_data]
-        except Exception as e:
-            raise LiteLLMClientError(
-                f"Batch embedding generation failed: {str(e)}"
-            ) from e
 
     def _build_completion_params(
         self, messages: list[dict[str, Any]], **kwargs: Any
